refactor: yarn & vue-directive-tooltip & audit fix
fix: because of issue with webpack and npm we have a package resolution issue with nuxt. So we have to move to yarn for now, this is the easiest way to resolve this issue (workaround). https://github.com/nuxt/nuxt.js/issues/4839
fix: security fix via npm audit --force
package change: move from markdown-loader@4.0.0 to markdown-loader@5.0.0 this major change in package version can lead to potentials error in code but this was required to pass the npm audit (resolved with npm audit --force)
fix: fix vue-directive-tooltip issue
package change: moved from vue-directive-tooltip to lefuturiste-vue-directive-tooltip because of a useless and not working postinstall script. Now on a forked package (follow the issue on the repository and be ready to revert to the original package). This can lead to potentials issue with the code later. https://github.com/hekigan/vue-directive-tooltip/issues/44